Added wide cameras + buttons + doors
Added wide cameras for puzzle sections. Also added doors that can be controlled by buttons. 2 types of buttons: 1 that just needs to be pushed, or needs to be pushed at the same time as its pair button, and 1 that needs to be held to keep the door open. Buttons are also scripted to be ready to be used for other interactable objects