Physics ball + gateway prefab + adjustments
Added a ball with appropriate physics such as gravity and collision. Also made a prefab containing the area blocks and camera transition that switches them. Finally, changed the way buttons work, to make them a more versatile tool. They are no longer limited to only 2, but can be used in conjunction with however many are needed. The doors have also been modified to work with this, as has the Interactables script